% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.vals}
\alias{get.vals}
\title{Get x and/or y values from a plot using the mouse}
\usage{
get.vals(axis = c("x", "y")[1], round.to = 1, print = TRUE)
}
\arguments{
  \item{axis}{Which axes to return values for, character
  vector}

  \item{round.to}{An integer number that values will be
  rounded to. For example if \code{round.to=100}, values
  will be rounded to the nearest 100.}

  \item{print}{Whether to print the values in the console,
  T/F}
}
\value{
A vector or matrix of the x and/or y values where the mouse
was clicked.
}
\description{
Get x and/or y values from a plot using the mouse
}
\details{
This function allows you to get the x and/or y values by
clicking on the active plot window. When you're done
clicking, right-click and select 'Stop' from the pop-up
menu.
}

