% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmap2.R
\name{gmap2}
\alias{gmap2}
\title{Download a static background image from Google}
\usage{
gmap2(x, exp = 1, type = "terrain", filename = "", style = NULL,
  scale = 1, zoom = NULL, size = c(640, 640), rgb = FALSE,
  lonlat = FALSE, google_apikey = NULL, ...)
}
\description{
Download a static background image from Google
}
