% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hm.expr.R
\name{hm.expr}
\alias{hm.expr}
\title{Hull metrics expressions}
\usage{
hm.expr(names.only = TRUE, desc = names.only, print = names.only)
}
\arguments{
\item{names.only}{Return only the names of hull metrics (not the expressions themselves). T/F.}

\item{desc}{Include descriptions of the hull metrics in the object returned. T/F.}

\item{print}{Display the list of hull metrics on the console. T/F.}
}
\value{
A named list with the following elements:
\enumerate{
\item ufat. User-friendly axis title
\item desc. Description (used to build plot captions)
\item def. Definition. If NULL, then the definition is the same as $desc
\item ufipt. User-friendly isopleth plot title. Used when plotting isopleths constructed by sorting hulls on the metric.
\item nhv. No hull value (the value to use for parent points that have no hulls, used to highlight the presence of points without hulls in histograms
\item iso.dec. Sort hulls in descending order when used to create isopleths). T/F
\item expr Expression object that when evaulated returns the hull metric values
\item zero2na Whether 0 values should be colored as NA (i.e., hidden) when classifying hull parent points, T/F
\item discrete Whether the metric assumes discrete (integer) values, used in classifying hull parent points T/F
\item auto.jiggle The maximum value of a uniform distribution used to apply a random offset to the metric when plotted on scatterplots to better see point density
\item spao.x include the metric as a default for the x-axis when the 'auto' option is used in \code{\link{lhs.plot.scatter.auto}}
\item spao.y include the metric as a default for the y-axis when the 'auto' option is used in \code{\link{lhs.plot.scatter.auto}}
\item req.metrics The underlying metric(s) that need to be saved in the hullset for successful computation. Character vector. This will usually simply be the name of the metric itself, but may be different in the case of derived metrics such as perimeter-area ratio.
\item req.ap The name(s) of required auxillary variables that are needed to extract the metric. Character vector.
\item req.ap.def A named list of default values of any required auxillary variables
\item req.ap.subtitle A named list of expression objects (one for each hmap) that returns a string of hmap values that will be used in the subtitle of plots
\item req.ap.desc An named list of expression objects (one for each hmap) that returns a string of hmap values that will be used in the description of plots
\item fun The name of the function that creates the hull metric
\item iso.hm2 Second hull metric used to break ties when creating isopleths (used mostly when the metric returns discrete values)
\item iso.hm2.dec Sort second hull metric descending
}
}
\description{
Returns a list of expression objects that when evaluated in functions will return various hull metrics
}
\details{
Hull metrics are used by many functions. This function returns a list of the expressions to 'pull out' hull metric values from
the data structure of a \code{\link{LoCoH-hullset}} object and other properties of hull metrics. These objects are used
by other T-LoCoH functions for plotting, exporting, creating isopleths, and other things you can do with hull metrics.

Most users will not need to use this function other than to see a list of hull metric names (which you need to know to specify a
hull metric in a plotting function for example) and the hull metric descriptions.

This function does not compute any hull metrics, only provide a comprehensive list of all possible metrics.
Some hull metrics are computed by default when you create a hullset (see \code{\link{lxy.lhs}}).
Others metrics must be computed with a special function (e.g., \code{\link{lhs.ellipses.add}}. To see which hull metrics exist for a specific LoCoH-hullset,
use \code{\link{summary.locoh.lhs}}.
}
\seealso{
\code{\link{lhs.ellipses.add}}, \code{\link{lhs.visit.add}}
}

