\name{hmap.grp}
\alias{hmap.grp}
\title{Returns a preset for a group of indices for hull metric auxillary parameters}
\usage{
  hmap.grp(n = c(3, 5, 7)[2], hlim = c(0.6667, 0.6667),
    slim = c(0.1225, 1), vlim = c(0.96, 1))
}
\description{
  Returns a preset for a group of indices for hull metric
  auxillary parameters
}

