% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp2expr.R
\name{hsp2expr}
\alias{hsp2expr}
\title{Returns a R expression that will recreate a saved scatterplot of hull metrics}
\usage{
hsp2expr(hsp, reg_nodes_prec = 3, var_name = "hsp")
}
\arguments{
\item{hsp}{A list of objects of class locoh.hsp}

\item{reg_nodes_prec}{The number of decimal places that will be preserved when printing the coordinates of the nodes of manually defined polygon regions in scatterplot space}

\item{var_name}{The name(s) of the variable that will be initialized in the output code}
}
\value{
A chunk of R code that will reproduce the \code{hsp} object
}
\description{
Returns a R expression that will recreate a saved scatterplot of hull metrics
}
\details{
Hull metric scatterplot settings can be saved as an object for reuse as a map legend or subsetting
data (see \emph{T-LoCoH Tips} on website). Saved scatterplot settings can include the boundaries of manually
defined (i.e., with the mouse) regions of interest, that can be used as a map legend or for data subsetting.
The purpose of this function is to generate R code so that 
polygons that were manually defined with the mouse in a hull metric scatterplot can be recreated
(e.g., in a markdown document) for the purposes of reproducible research. \code{hsp2expr} will
return a valid chunk of R code that recreates the saved scatterplot settings. This code can be
copy-pasted into a RMarkdown document or other report format so the work may be recreated.
}

