% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lhs.ellipses.add.R
\name{lhs.ellipses.add}
\alias{lhs.ellipses.add}
\title{Compute hull bounding ellipses}
\usage{
lhs.ellipses.add(lhs, id = NULL, k = NULL, r = NULL, a = NULL,
  s = NULL, hs.names = NULL, status = TRUE, beep = FALSE,
  save.ellipses = TRUE, existing.ellipses = c("overwrite", "abort")[1])
}
\arguments{
  \item{lhs}{A LoCoH-hullset object}

  \item{id}{The name(s) of individuals to analyze}

  \item{k}{The k value of hullsets to analyze}

  \item{r}{The r value of hullsets to analyze}

  \item{a}{The a value of hullsets to analyze}

  \item{s}{The s value of hullsets to analyze}

  \item{hs.names}{The name(s) of saved hullsets to analyze}

  \item{status}{Show status messages (T/F)}

  \item{beep}{Beep when done (T/F)}

  \item{save.ellipses}{Save ellipses in the lhs object?
  (T/F). If False, only the eccentricity values of the
  ellipses are saved as a hull metric.}

  \item{existing.ellipses}{Whether to overwrite or abort if
  existing ellipses are saved}
}
\value{
A LoCoH-hullset object
}
\description{
Creates bounding ellipses around each hull and computes the
eccentricity as a metric of ellongation
}
\note{
Saving the ellipses is optional. If ellipses are not saved,
the eccentricity values will still be computed and saved as
a hull metric. The only real reason to save ellipses is if
you wish to plot them.

Note creating ellipses is computationally intensive and can
take a long time.
}

