% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lhs.hsp.add.R
\name{lhs.hsp.add}
\alias{lhs.hsp.add}
\title{Adds a hull scatterplot to a LoCoH-hullset object}
\usage{
lhs.hsp.add(lhs, hsp.lst)
}
\arguments{
\item{lhs}{A LoCoH-hullset object}

\item{hsp.lst}{A list of object(s) of class locoh.hsp}
}
\description{
Adds a hull scatterplot to a LoCoH-hullset object
}
\note{
hsp objects are typically created with \code{\link{lhs.plot.scatter}} function. They can then be 'saved' in a LoCoH-hullset object with this function.
}
\examples{
## Display a scatter plot and manually draw four regions on it
## hsp <- lhs.plot.scatter(lhs, x="area", y="par", regions=4)

## Save hull scatter plot as part of the hullset
## lhs <- lhs.hsp.add(lhs, hsp.lst=hsp)
## summary(lhs)
}

