% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lhs.ids.get.R
\name{lhs.ids.get}
\alias{lhs.ids.get}
\title{Get the ids in a hullsets}
\usage{
lhs.ids.get(lhs, id = NULL, k = NULL, r = NULL, a = NULL, s = NULL,
  hs.names = NULL, hs.idx = NULL)
}
\arguments{
\item{lhs}{A \link{LoCoH-hullset} object}

\item{id}{The name(s) of individuals to extract}

\item{k}{The k value of hullsets to extract}

\item{r}{The r value of hullsets to extract}

\item{a}{The a value of hullsets to extract}

\item{s}{The s value of hullsets to extract}

\item{hs.names}{The name(s) of saved hullsets to extract}

\item{hs.idx}{The indices of saved hullsets to extract}
}
\value{
A character vector of ids
}
\description{
Get the ids in a hullsets
}
\details{
This function returns the ids of specified hullsets
}
\seealso{
\code{\link{lhs.select}}
}

