% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lxy.merge.R
\name{lxy.merge}
\alias{lxy.merge}
\title{Merges LoCoH-xy object}
\usage{
lxy.merge(..., pts.flds = "all", save.ptid = FALSE, new.id = NULL,
  save.old.id = NULL, dt.int.round.to = 0.1, tau.diff.max = 0.02)
}
\arguments{
\item{...}{\link{LoCoH-xy} objects}

\item{pts.flds}{A character vector of column names to include in the merged object, can also be \code{"all"} or \code{NULL}}

\item{save.ptid}{If duplicate ptid values are encounted in the merged object, whether to save the original ptid values as an ancillary variable, T/F}

\item{new.id}{A new id value for the merged object}

\item{save.old.id}{A boolean or the name of a ancillary variable where the old id values will be saved}

\item{dt.int.round.to}{The proportion of the median sampling frequency that time intervals will be rounded to when
computing the frequency table of sampling intervals (no change is made to the time stamps)}

\item{tau.diff.max}{The maximum deviation from tau (the median delta.t of the entire dataset), expressed as a proportion of tau, that
time difference between two points must fall for the distance between those two points to be included in the calculation of the median step length}
}
\description{
Merges LoCoH-xy objects together
}
\details{
The coordinate system of the LoCoH-xy objects being merged must be the same. Also each lxy object must have the same data columns
present, e.g., dt, col, ancillary variables, etc. To merge a subset of the attribute fields, use the \code{pts.flds} parameter.

Because LoCoH-xy object must have unique ptid values for each location, if the merged object contains duplicate ptid values
new one will be created. If \code{save.ptid=TRUE}, the 'old' ptid values will be saved as \code{ptid.orig}.

The merged object will no have nearest neighbors table and the random walk parameters will be recomputed.
}
\seealso{
\code{\link{lxy.id.new}}
}

