% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxy.plot.tspan.R
\name{lxy.plot.tspan}
\alias{lxy.plot.tspan}
\title{Plot distributions of the time span of nearest neighbors}
\usage{
lxy.plot.tspan(lxy, s = NULL, k = NULL, a = NULL, r = NULL, id = NULL,
  type = c("hist", "boxplot", "mean")[1], outline = FALSE, breaks = 20,
  desc = c(0, 1, 3)[2], cex.desc = 0.8, col.desc = "darkgreen",
  col.hist = "gray80", mar = c(if (type == "hist") 1.5 else 3, 3, if (type
  == "hist") 2 else if (title.show) 3 else 0.5, 0.5), mgp = c(1.9, 0.5, 0),
  figs.per.page = NULL, title = NULL, title.show = TRUE,
  title.obj.name = FALSE, panel.num = NULL,
  panel.num.inside.plot = !title.show, panel.num.cex = 2,
  no.sci.notation = FALSE, png.dir = NULL, png.dir.make = TRUE,
  png.width = 800, png.height = png.width, png.overwrite = TRUE,
  png.pointsize = 12 + (png.width - 480)/80, status = TRUE, ...)
}
\arguments{
\item{lxy}{A \link{LoCoH-xy} object}

\item{s}{The s value(s) of nearest neighbor sets to include in the plot. If NULL, all values will be used}

\item{k}{A k-value for the number of nearest neighbors around each point to include in the plot}

\item{a}{A a-value for the number of nearest neighbors around each point to include in the plot}

\item{r}{A r-value for the number of nearest neighbors around each point to include in the plot}

\item{id}{The id(s) of the individual(s) to include in the plot}

\item{type}{The type of plot for the time span: \code{'hist'}, \code{'boxplot'}, or \code{'mean'}}

\item{outline}{Show outliers in the box plots T/F}

\item{breaks}{The breaks parameter for a histogram, see \code{\link{hist}}}

\item{desc}{Which side to display automatically generated desciptive text (e.g. caption). 0=none, 1=bottom, 3=top.}

\item{cex.desc}{The expansion factor for the descriptive text. Numeric value.}

\item{col.desc}{The color of the descriptive text. Color value.}

\item{col.hist}{The color of the histogram bars. Color value.}

\item{mar}{The plot margins. A four item numeric vector}

\item{mgp}{The distance away from the edge of the plot for the 1) label, 2) tick marks, and 3) axis line. A three-item numeric vector}

\item{figs.per.page}{The number of plots per page.}

\item{title}{The title to be displayed. Character. If NULL a title will be constructed.}

\item{title.show}{Whether to show the title. T/F}

\item{title.obj.name}{Whether to add the name of the lxy object to the plot title (ignored if \code{title} is passed). T/F}

\item{panel.num}{A number or letter to display in the upper left hand corner of the plot when the plot will be used as part of a multi-frame graphic (as in publications). Character}

\item{panel.num.inside.plot}{Whether to display panel.num inside the plot area itself, as opposed to the title area. Ignored if panel.num is NULL. T/F}

\item{panel.num.cex}{The expansion factor for the panel number. Ignored if panel.num is NULL.}

\item{no.sci.notation}{Whether to avoid the use of scientific notation on labels on the x-axis. T/F}

\item{png.dir}{The directory for a PNG file (filename will be constructed automatically). Ignored if png.fn is passed}

\item{png.dir.make}{Whether to create png.dir if it doesn't exist. T/F}

\item{png.width}{The width of the PNG image}

\item{png.height}{The height of the PNG image}

\item{png.overwrite}{Whether to overwrite an existing PNG file if it exists. T/F}

\item{png.pointsize}{The pointsize (in pixels) for the PNG image, equivalent to the height or width of a character in pixels (increase to make labels appear larger)}

\item{status}{Show progress bar and status messages. T/F}

\item{...}{Additional parameters that will be passed to the \code{\link{plot}} function}
}
\value{
A list of plots created, with one element per id, and each element consisting of another 
list with elements for the filename (or NULL), filename dimensions in pixels (or NULL), the timespan values 
(as a list with one element for each value of s), descriptive text, and the k/a/r value
}
\description{
Plot distributions of the time span of nearest neighbors as a proportion of the median sampling interval
}

