\name{strTrim}
\alias{strTrim}
\title{Trim leading and/or trailing blanks from a character object}
\usage{
strTrim(str, side = "both")
}
\arguments{
  \item{str}{Input string}

  \item{side}{Which side(s) of the input string to trim:
  'left', 'right', or 'both'}
}
\description{
Trim leading and/or trailing blanks from a character object
}

