\name{summary.locoh.lhs}
\alias{summary.locoh.lhs}
\title{summary.locoh.lhs}
\usage{
  summary.locoh.lhs(lhs, file = "", id = NULL, k = NULL,
    r = NULL, a = NULL, s = NULL, hs.names = NULL,
    iso.details = FALSE, desc = FALSE, compact = FALSE)
}
\arguments{
  \item{lhs}{A LoCoH-hullset object}

  \item{file}{A file name}

  \item{id}{The name(s) of individuals to summarize}

  \item{k}{The k value of hullsets to summarize}

  \item{r}{The r value of hullsets to summarize}

  \item{a}{The a value of hullsets to summarize}

  \item{s}{The s value of hullsets to summarize}
}
\description{
  Prints a summary of a LoCoH-hullset object
}

