\name{summary.locoh.lxy}
\alias{summary.locoh.lxy}
\title{Show a summary of a LoCoH-xy object}
\usage{
  summary.locoh.lxy(lxy, file = "", dt.int = FALSE,
    round.coords = 1, ptsh = FALSE)
}
\arguments{
  \item{lxy}{\code{\link{LoCoH-xy object}}}

  \item{file}{A file name where the results will be saved}

  \item{dt.int}{Whether to show a summary of the sampling
  interval}
}
\description{
  Prints a summary of a locoh xy object (set of locations)
}
\seealso{
  \code{\link{lxy.plot.freq}}
}

