\name{text.contrast.color}
\alias{text.contrast.color}
\title{Choose a white or black plot background for a particular color of text}
\usage{
  text.contrast.color(x, ...)
}
\arguments{
  \item{x}{A color value}

  \item{...}{Other arguments (not used)}
}
\description{
  Choose a white or black plot background for a particular
  color of text
}

