% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textContrastColor.R
\name{textContrastColor}
\alias{textContrastColor}
\title{Choose a white or black plot background for a particular color of text}
\usage{
textContrastColor(x, ...)
}
\arguments{
\item{x}{A color value}

\item{...}{Other arguments (not used)}
}
\description{
Choose a white or black plot background for a particular color of text
}
