\name{tsd.zvals}
\alias{tsd.zvals}
\title{tsd.zvals}
\usage{
  tsd.zvals(delta.t, sVal, type, d.bar = NULL, tau = NULL,
    vmax = NULL)
}
\arguments{
  \item{delta.t}{is the time difference in seconds}

  \item{sVal}{is the value of S}

  \item{d.bar}{is the median step length (for the entire
  dataset)}

  \item{tau}{is the median sampling interval (for the
  entire dataset)}
}
\description{
  returns the values for the "time axis" of the TSD
  distance metric
}

