% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorize.parameter.R
\name{vectorize.parameter}
\alias{vectorize.parameter}
\title{Convert a comma-separated string to a vector}
\usage{
vectorize.parameter(x, n2z = FALSE, type = c("numeric", "character")[1],
  sort.res = TRUE)
}
\arguments{
\item{x}{The vector or character string to be evaluated}

\item{n2z}{Convert null values to zero or an empty string (depending on \code{type}}

\item{type}{The data type: \code{numeric} or \code{character}}

\item{sort.res}{Whether to sort the result. T/F}
}
\value{
A vector
}
\description{
Convert a comma-separated string to a vector
}

