% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_slivers.R
\name{clean_slivers}
\alias{clean_slivers}
\title{Clean slivers}
\usage{
clean_slivers(sp, status = TRUE, min_nodes = 3)
}
\arguments{
\item{sp}{A SpatialPolygons or SpatialPolygonsDataFrame object}

\item{status}{Show messages. T/F}

\item{min_nodes}{The minimum number of unique nodes a polygon must have to *not* be labeled a sliver and deleted.}
}
\value{
Returns a list with two elements:

  sp - the (cleaned) SpatialPolygons* object
  results - a dataframe of the polygon and Polygon indices of removed slivers
}
\description{
Remove 'slivers' from a SpatialPolygons* object
}
\details{
As defined here, a 'sliver' is polygon or hole that has < \code{min_nodes} (default=3) unique nodes
(perhaps due to a data type conversion or rounding error). 
This seems to be a particularly common type of geometry error that 
crops in when unioning hulls into isopleths.
}
