% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_slivers.R
\name{clean_slivers}
\alias{clean_slivers}
\title{Clean slivers}
\usage{
clean_slivers(sp, status = TRUE)
}
\description{
This will remove 'slivers' from a SpatialPolygons* object.
As defined here, a 'sliver' is polygon or hole that has < 3 unique nodes
(perhaps due to a data type conversion or rounding error).
This seems to be a particularly common type of geometry error that
crops in when merging hulls into isoleths.
Returns a list with two elements
  sp - the (cleaned) SpatialPolygons* object
  results - a dataframe of the polygon and Polygon indices of removed slivers
}

