\name{iso2raster}
\alias{iso2raster}
\title{Converts a SpatialPolygonsDataFrame to a raster
This presumes the SPDF contains isopleths ordered from lowest level to highest}
\usage{
  iso2raster(polys, ext = NULL, cell.size = NULL,
    ll.round = TRUE, status = TRUE)
}
\description{
  Converts a SpatialPolygonsDataFrame to a raster This
  presumes the SPDF contains isopleths ordered from lowest
  level to highest
}

