% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lhs.anv.add.R
\name{lhs.anv.add}
\alias{lhs.anv.add}
\title{Adds an ancillary value to a hullset}
\usage{
lhs.anv.add(lhs, id = NULL, k = NULL, r = NULL, a = NULL, s = NULL,
  hs.names = NULL, anv, anv.desc = NULL, overwrite = FALSE)
}
\arguments{
\item{lhs}{A LoCoH-hullset object}

\item{id}{The name(s) of individuals to export}

\item{k}{The k value of hullsets to export}

\item{r}{The r value of hullsets to export}

\item{a}{The a value of hullsets to export}

\item{s}{The s value of hullsets to export}

\item{hs.names}{The name(s) of saved hullsets to export}

\item{anv}{A vector, named list, or data frame with the same number of values as points}

\item{anv.desc}{A character vector of descriptions}

\item{overwrite}{Whether to overwrite existing variables with the same names (T/F)}
}
\value{
A \link{LoCoH-hullset} object
}
\description{
Add ancillary values to the points of a hullset, which can then be used as hull metrics
}
\details{
This will add one or more ancillary values (columns) to the point locations
in a \link{LoCoH-hullset} object.
}

