% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.filter.hsp.R
\name{lhs.filter.hsp}
\alias{lhs.filter.hsp}
\title{Define subsets of hulls based on the location of the parent point in hull scatterplot space}
\usage{
lhs.filter.hsp(lhs, id = NULL, k = NULL, r = NULL, a = NULL, s = NULL,
  hs.names = NULL, hsp = NULL, reg.idx = NULL, label = NULL,
  col = NULL)
}
\arguments{
\item{lhs}{A LoCoH-hullset object}

\item{id}{The name(s) of the individual(s) to include in the plot. Character vector or comma-delimited character.}

\item{k}{The k value of hullsets to create isopleths for}

\item{r}{The r value of hullsets to create isopleths for}

\item{a}{The a value of hullsets to create isopleths for}

\item{s}{The s value of hullsets to create isopleths for}

\item{hs.names}{The name(s) of saved hullsets to create isopleths for}

\item{hsp}{Either the index of a hull scatterplot(s) saved in lhs (use \code{\link{summary.locoh.lhs}} to which how many hsp 
objects have been saved), or a list of objects of class \code{locoh.hsp} (i.e., the return value of \code{\link{lhs.plot.scatter}}.}

\item{reg.idx}{A numeric vector of the indices of the regions in \code{hsp} to include in the filter}

\item{label}{Character vector of the labels to use for each subset. If not passed, the saved label values in \code{hsp} will be used}

\item{col}{Vector of color values (one per region). If omitted the colors saved in \code{hsp} will be used}
}
\value{
A list that defines subsets of hulls, with one element for each region in \code{hsp}. Each element is a list with three elements:
\describe{
   \item{idx}{indices of the hull parent points}
   \item{label}{the label of the region of hsp}
   \item{col}{the color of the region in hsp}
}
}
\description{
Define subsets of hulls based on the location of the parent point in hull scatterplot space
}
\details{
This will return a list that defines subsets of hulls grouped according to which manually-digitized region in scatterplot 
space the hull parent point falls. This can be passed to several functions, including \code{\link{plot.locoh.lhs}} (future), 
and \code{\link{lhs.plot.scatter}}, to create plots of subsets of hulls.

The scatterplot space, including the manually defined regions, are saved in a \code{locoh.hsp} object. Typically \code{hsp} will be created
using the \code{\link{lhs.plot.scatter}} function. \code{hsp} objects can also be saved in the \code{lhs} object (see \code{\link{lhs.hsp.add}}.

Note that this function can only return subsets for *one* hullset and *one* hsp object. This means that either \code{lhs} must contain a single
hullset, or other parameters (e.g., \code{id}, \code{hs.names}, \code{k}, \code{r}, and/or \code{a}) are passed to select one and only one of the hullsets in \code{lhs}.
}

