% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lhs.save.R
\name{lhs.save}
\alias{lhs.save}
\title{Save a LoCoH-hullset object to disk}
\usage{
lhs.save(lhs, file = NULL, dir = ".", suf = NULL, compress = TRUE,
  auto.num.files = TRUE, width = 2, save.as = NULL)
}
\arguments{
  \item{lhs}{A LoCoH-hullset object}

  \item{file}{Optional filename. If omitted a sensible
  filename will be constructed}

  \item{dir}{Directory for the output file (relative to the
  working directory), ignored if \code{file} passed}

  \item{suf}{A suffix that will be used in the construction
  of the filename, ignored if \code{file} is passed}

  \item{compress}{Compress file. T/F}

  \item{auto.num.files}{Use a number as part of the
  constructed filename. T/F}

  \item{width}{The number of digits of the auto-number
  token in the filename (ignored if
  \code{auto.num.files=F})}

  \item{save.as}{The name of the object when saved (default
  is the same as the original)}
}
\description{
Saves a LoCoH-hullset object to disk, constructing a
filename if needed that reflects the contents
}
\seealso{
\code{\link{lxy.save}}
}

