% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.select.R
\name{lhs.select}
\alias{lhs.select}
\title{Select hullsets}
\usage{
lhs.select(lhs, id = NULL, k = NULL, r = NULL, a = NULL, s = NULL,
  hs.names = NULL, hs.idx = NULL, selection = NULL, status = TRUE)
}
\arguments{
\item{lhs}{A \link{LoCoH-hullset} object}

\item{id}{The name(s) of individuals to extract}

\item{k}{The k value of hullsets to extract}

\item{r}{The r value of hullsets to extract}

\item{a}{The a value of hullsets to extract}

\item{s}{The s value of hullsets to extract}

\item{hs.names}{The name(s) of saved hullsets to extract}

\item{hs.idx}{The indices of saved hullsets to extract}

\item{selection}{A locoh.selection object, see Details}

\item{status}{Show status messages (T/F)}
}
\value{
A LoCoH-hullset object
}
\description{
Select hullsets
}
\details{
This function returns a subset of hullsets based on the id value, k, a, r, or s. You may also select hullsets based 
on hullset names, indices, or a hullset selection object. A \emph{locoh.selection} object is 
a named list, with one element for each id, containing one or more a, k, or r value(s). 

See the \code{lhs.selection} function in the tlocoh.dev package for details.
}

