% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxy.gridanv.add.R
\name{lxy.gridanv.add}
\alias{lxy.gridanv.add}
\title{Add ancillary variable(s) to a LoCoH-xy object from a raster}
\usage{
lxy.gridanv.add(lxy, band = NULL, dtfn = NULL, fn = NULL,
  anv.name = NULL, anv.desc = NULL, date.match = c("closest",
  "before")[1])
}
\arguments{
\item{lxy}{\link{LoCoH-xy} object}

\item{band}{Numeric vector of which band(s) to read. If omitted, all bands will be read.}

\item{dtfn}{A two-column data frame containing a date and filename}

\item{fn}{Filename of a raster}

\item{anv.name}{The name(s) of the ancillary variable(s) that will be added to the LoCoH-xy object (one per band)}

\item{anv.desc}{A character vector of short descriptions of each ancillary variable}

\item{date.match}{How to match dates when \code{dtfn} is used: \code{'closest'} or \code{'before'}. See note.}
}
\value{
A \link{LoCoH-xy} object
}
\description{
Add ancillary variable(s) to a LoCoH-xy object from a raster
}
\note{
This function will create new ancillary variables and fill it with the value of the pixel at each point.
The image should be in the same coordinate system as the points (the script however does not check this).
Any points that fall beyond the edges of the image will be given 'NA' values. If an ancillary variable with the
same name already exists, it will be overwritten.

The raster image specified by \code{fn} must be in a format supported by the
\code{readGDAL} function in the rgdal package (run \code{gdalDrivers} for a list of supported formats).
Multi-band formats are supported. If the image has multiple bands, you can
specify which band(s) to read using the optional argument \code{band}. A
separate ancillary variable will be created for each band. You can provide
name(s) for the ancillary variable(s) with the \code{anv.name} parameter.

You can also pass a time-series of images, and the function will use the image from the closest date
for each point. To do this, pass a two-column data frame as the argument \code{dtfn}. The first
column of \code{dtfn} must be a date (in one of the POSIXt date-time classes), and the second column
must be a valid raster filename associated with that date. If each image is a time composite, (e.g.,
two-week NDVI), the date should either be the mid-point or the beginning date. Two options for time
matching are available through the argument \code{date.match}. When \code{date.match='closest'}, the
image whose date is closest in time (before or after) will be used. When \code{date.match='before'},
it will pick the first image that comes before the location.
}
