\name{textContrastColor}
\alias{textContrastColor}
\title{Choose a white or black plot background for a particular color of text}
\usage{
  textContrastColor(x, ...)
}
\arguments{
  \item{x}{A color value}

  \item{...}{Other arguments (not used)}
}
\description{
  Choose a white or black plot background for a particular
  color of text
}

